requirements.txt
streamlit
pandas
replicate
python-dotenv
openai